implement inotify-based automatic reload
There is no generic interface for this yet in the SDK or the toolkit as
a whole, but before implementing a poller system I can just as well go
for my 'usual' pattern of using inotifywait for the heavy lifting.

That isn't exactly nice from a porter perspective, but all my systems
are Linux have inotify available, so this can surf me well until a
generic interface is available.