allow scripts to show desktop notifications
A website can display a desktop notification via the WebAPI, but it has
to ask for permission first. Scripts on the other hand are written by
the user, so if the user implemented a notification its reasonable to
assume that this also implicitly grants permission to show it. It
doesn't grant the (modified) website the same rights through, so instead
of using the WebAPI we introduce here a simple port-communication to
show a notification (unconditionally) via the SDK.