allow content scripts to spawn shell scripts
Allowing content scripts to talk to binaries seems like overkill at
first and it might is in various instances, but if it is strictly
controlled it can be handy e.g. to call external downloaders or
spice up the resource://dotpagemod-config/ interface with buttons to
create and delete the configuration from inside firefox and
by opening files to edit in your preferred external editor.

The limiting choice is here that a content script can't call any binary
as it pleases, but can only call scripts in its own host directory –
which not only surfs to controll what can be called but makes it also
blantly obvious that these scripts can make such calls.