just reload the effected pagemod if a file was modified
Modifying a single file is a common task so avoiding work is a good
strategy, which is just a tiny bit complicated by how e.g. vim is
implementing editing files via move (to backup), create file and insert
content – but no pain no gain I guess.

We need to reload the pagemod only on create/delete as modification is
detected by firefox – but we want to reload the pagemod on modify anyhow
to have the changes be automatically applied rather than requiring the
user to reload the page.