add examples: remove external page tracking
Beside that I don't see a reason to tell those sites to which sites I
navigate to, the also deploy antifeatures like intermediate landing
pages telling me that I leave their walled garden and how insecure that
might be, the link target is harder to read (they all hack around this
with javascript code on their own, but I don't run their code…) and so
on.

The code itself is hardly topnodge and e.g. doesn't revert on unapply,
but such trivial page mods are the reason why this extension exists at
all, so more examples are always good.