document how to (not) work with images
Making files available on a fixed path like before isn't really possible
anymore and had its drawbacks anyhow, so as sad as it might be we take
the easy way out and go with embedding files for now. I wasn't really
using it much anyhow to warrant a more complicated solution for now.