wrap individual scripts in function to avoid leaks
All scripts for a tab run in the same sandbox which could potentially
lead to scripts running into conflicts if they accidently share some
variables. Such a leak is of course intended for FRAMEWORK files, but
for the rest we can do at least a little better by wrapping the scripts
in individual anonymous functions.