a failed file isn't already in the tab
If a tab was opened in a previous session and now regains focus firefox
will load the tab and for some reason does trigger the load event for us
two times: Once to early most of the time and a second time later at a
point the applying will work. To workaround this for now we consider
files which failed in applying as not being applied for this tab yet
causing them to be applied (again).