app: watch symlinked directories explicitly
inotifywait watches for changes on the symlink, it isn't following the
symlink, so we not only watch the config directory but also every
"subdirectory" which is a symlink to get notifications about changes in
them as well. Important if you want to share some packs between profiles
or e.g. pull in the examples/ from this repository into your own config.

New directories symlinked to aren't picked up automatically, so you will
have to restart the app manually (button in popup) for now in case you
add such symlinks.