examples: zeit.de: show images without js and noscript
They have <noscript> tags to show the images in case javascript is
disabled, which pushes them way ahead of the crowd – but I don't have
javascript disabled, so it doesn't work in my setup (not their fault!).
As it is easy enough to add the image "by hand" I do that instead of
inserting the whole <noscript> content as HTML into the page.