add a test for URI to host directories
Jasmine is used to implement the test, but there is no infrastructure
added for it to be run at the moment until I (or someone else) figure
out a sensible way of making it run non-interactively.

For now its more a matter of documenting in code how this mapping works
which by itself is a good thing already.