make xpi building slightly reproducible
zip files contain timestamps and the xpi contains generated files,
so the built xpi never matches earlier builds even if the content
is identical.

Based on the tips collected by the Reproducible Build team:
https://wiki.debian.org/ReproducibleBuilds/TimestampsInZip

Note that this only works if operated in the same checkout for now
which drastically hampers the applicability in reality.