Apply pagemods to sites with ports consistently
The initial load would apply files in 'example.org' if you visit
'example.org:8080', but a reload wouldn't reapply files from the later
directory. For consistency and to support both possible choices on that
spectrum 'example.org' files will be applied to visits and reloads
regardless of a portnumber present or not. If you want files to apply
only to hostnames without a port you can use a portspecific directory
with the portnumber being the default for the used protocol.

Test is adapted and the README has a cheatsheet entry now about the
possible naming of directories (not) involving ports. Its not THAT
common in practice as it took me 2 years to notice…