Allow scripts to choose readyState by filename
It isn't common, but sometimes `interactive` (aka `document_end`) is too
late/early for a script to run, so for those cases it is handy to choose
a different insertion point.

An example are scripts using MutationObservers. Another is providing a
FRAMEWORK-like script only for certain hostnames.

The same is technically allowed and supported for CSS as well, but I
don't see how that could be useful at the moment.