Show script failure message in console and panel
Failed scripts are highlighted in red to indicate that something went
wrong, but finding out what was wrong was a bit too hard. We solve this
by printing the error to the error console of the page we added to
(note while that shows the correct filepath, clicking it will bring you
 just to the error message generation instead of the real file).

We also show the error message on hover over the filename in the panel
which has the advantage of being clickable and opening the correct file
(assuming a working editor was configured in the host app of course).