Avoid confusing Firefox's version parsing with git describe
If you have a git checkout with commits above the last tag we generated
a version like "v0.4.1-21-ga49423e", but that confuses Firefox as it
seems to assume we would be using semversion and spams the browser
console with "Reading manifest: Warning processing version: version must
be a version string consisting of at most 4 integers of at most 9 digits
without leading zeros, and separated with dots". So, we strip down the
example version down to "v0.4.1.21" for Firefox to be happy.